+++ packages/cdai/src/components/IdeSlideOverPanel/IdeSlideOverPanel.js
@@ -56,7 +56,8 @@ const IdeSlideOverPanel = ({
           className={`${idePrefix}-slide-over-panel--secondary`}
           disabled={secondaryButtonDisabled}
           renderIcon={secondaryButtonIcon}
-          onClick={onSecondaryClick}>
+          onClick={onSecondaryClick}
+        >
           {secondaryButtonText}
         </Button>
         <Button
@@ -64,7 +65,8 @@ const IdeSlideOverPanel = ({
           kind={danger ? 'danger' : 'primary'}
           disabled={primaryButtonDisabled}
           renderIcon={primaryButtonIcon}
-          onClick={onPrimaryClick}>
+          onClick={onPrimaryClick}
+        >
           {primaryButtonText}
         </Button>
       </div>
@@ -125,6 +127,7 @@ const IdeSlideOverPanel = ({
 };
 
 IdeSlideOverPanel.propTypes = {
+  children: PropTypes.node,
   className: PropTypes.string,
   closeButtonIconDescription: PropTypes.string,
   controls: PropTypes.bool,