+++ packages/cloud-cognitive/src/components/ImportModal/ImportModal.js
@@ -167,7 +167,8 @@ export let ImportModal = forwardRef(
         aria-label={title}
         className={cx(blockClass, className)}
         size="sm"
-        preventCloseOnClickOutside>
+        preventCloseOnClickOutside
+      >
         <ModalHeader className={`${blockClass}__header`} title={title} />
         <ModalBody className={`${blockClass}__body-container`}>
           {description && (
@@ -199,7 +200,8 @@ export let ImportModal = forwardRef(
               onClick={fetchFile}
               className={`${blockClass}__import-button`}
               size="sm"
-              disabled={importButtonDisabled}>
+              disabled={importButtonDisabled}
+            >
               {inputButtonText}
             </Button>
           </div>
@@ -232,7 +234,8 @@ export let ImportModal = forwardRef(
             type="submit"
             kind="primary"
             onClick={onSubmitHandler}
-            disabled={primaryButtonDisabled}>
+            disabled={primaryButtonDisabled}
+          >
             {primaryButtonText}
           </Button>
         </ModalFooter>