+++ packages/cloud-cognitive/src/components/SidePanel/SidePanel.js
@@ -527,7 +527,8 @@ export let SidePanel = React.forwardRef(
             [`${blockClass}__title-container-is-animating`]:
               !animationComplete || !open,
             [`${blockClass}__title-container-without-title`]: !title,
-          })}>
+          })}
+        >
           {currentStep > 0 && (
             <Button
               aria-label={navigationBackIconDescription}
@@ -565,7 +566,8 @@ export let SidePanel = React.forwardRef(
               [`${blockClass}__subtitle-text-is-animating`]:
                 !animationComplete && animateTitle,
               [`${blockClass}__subtitle-without-title`]: !title,
-            })}>
+            })}
+          >
             {subtitle}
           </p>
         )}
@@ -573,7 +575,8 @@ export let SidePanel = React.forwardRef(
           <div
             className={cx(`${blockClass}__action-toolbar`, {
               [`${blockClass}__action-toolbar-no-animation`]: !animateTitle,
-            })}>
+            })}
+          >
             {actionToolbarButtons.map((action) => (
               <Button
                 key={action.label}
@@ -600,7 +603,8 @@ export let SidePanel = React.forwardRef(
                     ? action.onClick(event)
                     : action.onActionToolbarButtonClick &&
                       action.onActionToolbarButtonClick(event)
-                }>
+                }
+              >
                 {action.leading && action.label}
               </Button>
             ))}
@@ -615,7 +619,8 @@ export let SidePanel = React.forwardRef(
           <h2
             className={`${blockClass}__title-text`}
             title={title}
-            aria-hidden={false}>
+            aria-hidden={false}
+          >
             {title}
           </h2>
         )}
@@ -623,7 +628,8 @@ export let SidePanel = React.forwardRef(
           <h2
             className={`${blockClass}__collapsed-title-text`}
             title={title}
-            aria-hidden={true}>
+            aria-hidden={true}
+          >
             {title}
           </h2>
         )}
@@ -662,12 +668,14 @@ export let SidePanel = React.forwardRef(
             onBlur={handleBlur}
             ref={contentRef}
             role="complementary"
-            aria-label={title}>
+            aria-label={title}
+          >
             <span
               ref={startTrapRef}
               tabIndex="0"
               role="link"
-              className={`${blockClass}__visually-hidden`}>
+              className={`${blockClass}__visually-hidden`}
+            >
               Focus sentinel
             </span>
             {!animateTitle && renderHeader()}
@@ -677,7 +685,8 @@ export let SidePanel = React.forwardRef(
                 [`${blockClass}__static-inner-content`]: !animateTitle,
                 [`${blockClass}__inner-content-with-actions`]:
                   actions && actions.length,
-              })}>
+              })}
+            >
               {animateTitle && renderHeader()}
               <div className={`${blockClass}__body-content`}>{children}</div>
               <ActionSet
@@ -690,7 +699,8 @@ export let SidePanel = React.forwardRef(
               ref={endTrapRef}
               tabIndex="0"
               role="link"
-              className={`${blockClass}__visually-hidden`}>
+              className={`${blockClass}__visually-hidden`}
+            >
               Focus sentinel
             </span>
           </div>